refactor: dsearch apis
