+++ packages/web3-providers/src/CoinMarketCap/apis/DSearchAPI.ts
@@ -1,8 +1,8 @@
 import urlcat from 'urlcat'
 import type { SourceType, FungibleTokenResult } from '@masknet/web3-shared-base'
-import { fetchJSON } from '../../helpers/fetchJSON.js'
 import type { DSearchBaseAPI } from '../../types/DSearch.js'
 import { DSEARCH_BASE_URL } from '../../DSearch/constants.js'
+import { fetchFromDSearch } from '../../DSearch/helpers.js'
 
 export interface FungibleToken {
     id: string | number
@@ -21,7 +21,9 @@ export class CoinMarketCapSearchAPI<ChainId, SchemaType>
     implements DSearchBaseAPI.DataSourceProvider<ChainId, SchemaType>
 {
     async get(): Promise<Array<FungibleTokenResult<ChainId, SchemaType>>> {
-        const tokensURL = urlcat(DSEARCH_BASE_URL, '/fungible-tokens/coinmarketcap.json')
-        return fetchJSON<Array<FungibleTokenResult<ChainId, SchemaType>>>(tokensURL, { mode: 'cors' })
+        return fetchFromDSearch<Array<FungibleTokenResult<ChainId, SchemaType>>>(
+            urlcat(DSEARCH_BASE_URL, '/fungible-tokens/coinmarketcap.json'),
+            { mode: 'cors' },
+        )
     }
 }
