+++ packages/web3-providers/src/DSearch/rules.ts
@@ -1,6 +1,6 @@
 import { compact } from 'lodash-es'
 import { isSameAddress, type SearchResult, SearchResultType } from '@masknet/web3-shared-base'
-import type { Handler } from './type.js'
+import type { Handler } from './types.js'
 import { FuseAPI } from '../Fuse/index.js'
 
 const Fuse = new FuseAPI()
