+++ packages/web3-providers/src/NFTScan/apis/DSearchAPI.ts
@@ -1,8 +1,8 @@
 import urlcat from 'urlcat'
 import type { NonFungibleCollectionResult, SearchResult, SourceType } from '@masknet/web3-shared-base'
-import { fetchJSON } from '../../helpers/fetchJSON.js'
 import type { DSearchBaseAPI } from '../../types/DSearch.js'
 import { DSEARCH_BASE_URL } from '../../DSearch/constants.js'
+import { fetchFromDSearch } from '../../DSearch/helpers.js'
 
 export interface FungibleToken {
     id: string | number
@@ -19,11 +19,13 @@ export interface NonFungibleToken {
 
 export class NFTScanSearchAPI<ChainId, SchemaType> implements DSearchBaseAPI.DataSourceProvider<ChainId, SchemaType> {
     async get(): Promise<Array<SearchResult<ChainId, SchemaType>>> {
-        const nftsURL = urlcat(DSEARCH_BASE_URL, '/non-fungible-tokens/nftscan.json', { mode: 'cors' })
-        const collectionsURL = urlcat(DSEARCH_BASE_URL, '/non-fungible-collections/nftscan.json', { mode: 'cors' })
-        const nfts = fetchJSON<Array<SearchResult<ChainId, SchemaType>>>(nftsURL)
-        const collections = fetchJSON<Array<SearchResult<ChainId, SchemaType>>>(collectionsURL)
-        const collectionsFromSpecialList = await fetchJSON<Array<SearchResult<ChainId, SchemaType>>>(
+        const nfts = fetchFromDSearch<Array<SearchResult<ChainId, SchemaType>>>(
+            urlcat(DSEARCH_BASE_URL, '/non-fungible-tokens/nftscan.json', { mode: 'cors' }),
+        )
+        const collections = fetchFromDSearch<Array<SearchResult<ChainId, SchemaType>>>(
+            urlcat(DSEARCH_BASE_URL, '/non-fungible-collections/nftscan.json', { mode: 'cors' }),
+        )
+        const collectionsFromSpecialList = fetchFromDSearch<Array<SearchResult<ChainId, SchemaType>>>(
             urlcat(DSEARCH_BASE_URL, '/non-fungible-collections/specific-list.json', { mode: 'cors' }),
         )
         return (await Promise.allSettled([collectionsFromSpecialList, nfts, collections])).flatMap((v) =>
@@ -36,11 +38,12 @@ export class NFTScanCollectionSearchAPI<ChainId, SchemaType>
     implements DSearchBaseAPI.DataSourceProvider<ChainId, SchemaType>
 {
     async get(): Promise<Array<NonFungibleCollectionResult<ChainId, SchemaType>>> {
-        const collectionsURL = urlcat(DSEARCH_BASE_URL, '/non-fungible-collections/nftscan.json', { mode: 'cors' })
-        const collectionsFromSpecialList = await fetchJSON<Array<NonFungibleCollectionResult<ChainId, SchemaType>>>(
+        const collectionsFromSpecialList = fetchFromDSearch<Array<NonFungibleCollectionResult<ChainId, SchemaType>>>(
             urlcat(DSEARCH_BASE_URL, '/non-fungible-collections/specific-list.json', { mode: 'cors' }),
         )
-        const collections = await fetchJSON<Array<NonFungibleCollectionResult<ChainId, SchemaType>>>(collectionsURL)
+        const collections = fetchFromDSearch<Array<NonFungibleCollectionResult<ChainId, SchemaType>>>(
+            urlcat(DSEARCH_BASE_URL, '/non-fungible-collections/nftscan.json', { mode: 'cors' }),
+        )
         return (await Promise.allSettled([collectionsFromSpecialList, collections])).flatMap((v) =>
             v.status === 'fulfilled' && v.value ? v.value : [],
         )
